refactor(test): move human_bytes to utils
